+++ apps/admin/src/components/club/club-table.tsx
@@ -8,7 +8,7 @@ import {
 import { useClubServiceGetApiV1Clubs } from '~/apis/community/queries';
 import type { ClubDetailResponse } from '~/apis/community/requests';
 import useEditTable from '~/hooks/use-edit-table';
-import ClubCreate from './club-create';
+import ClubCreate from '~/components/club/club-creator';
 
 import ClubTableView from './club-table-view';
